requirements.txt
appdirs==1.4.4
blinker==1.9.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
curses-menu==0.9.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
prompt-tool-kit==1.0.14
Pygments==2.19.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
setuptools==80.9.0
six==1.17.0
urllib3==2.5.0
wcwidth==0.2.13
